refactor router

+++ server/resources/api/apiRouter.js
@@ -6,46 +6,39 @@ var apiController = require('./apiController');
 // =====================================================================
 apiRouter.route('/venues')
   .get(function(req, res) {
-    apiController.retrieve(req, res);
+    if (req.params.number) apiController.retrieveOne(req, res);
+    else apiController.retrieve(req, res);
   })
   .post(function(req, res) {
     apiController.createOne(req, res);
-  })
-  .delete(function(req, res) {
-    apiController.delete(req, res);
-  });
-
-
-apiRouter.route('/venues/:number')
-  .get(function(req, res) {
-    apiController.retrieveOne(req, res);
   })
   .put(function(req, res) {
     apiController.updateOne(req, res);
   })
   .delete(function(req, res) {
-    apiController.deleteOne(req, res);
+    if (req.params.number) apiController.deleteOne(req, res);
+    else apiController.delete(req, res);
+  });
+
+apiRouter.route('/keys')
+  .get(function(req, res) {
+    apiController.retrieveKey(req, res);
   });
 
 // apiRouter.route('/users')
 //   .get(function(req, res) {
-//     apiController.retrieve(req, res);
+//     if (req.params.number) apiController.retrieveOne(req, res);
+//     else apiController.retrieveOne(req, res);
 //   })
 //   .post(function(req, res) {
 //     apiController.createOne(req, res);
-//   })
-//   .delete(function(req, res) {
-//     apiController.delete(req, res);
-
-// apiRouter.route('/users/:number')
-//   .get(function(req, res) {
-//     apiController.retrieveOne(req, res);
 //   })
 //   .put(function(req, res) {
 //     apiController.updateOne(req, res);
 //   })
 //   .delete(function(req, res) {
-//     apiController.deleteOne(req, res);
+//     if (req.params.number) apiController.deleteOne(req, res);
+//     else apiController.delete(req, res);
 //   });
 
 module.exports = apiRouter;
